move validation to service
