+++ apps/server-asset-sg/src/app.module.ts
@@ -11,6 +11,7 @@ import { JwtMiddleware } from '@/core/middleware/jwt.middleware';
 import { PrismaService } from '@/core/prisma.service';
 import { AssetEditController } from '@/features/asset-edit/asset-edit.controller';
 import { AssetEditRepo } from '@/features/asset-edit/asset-edit.repo';
+import { AssetEditService } from '@/features/asset-edit/asset-edit.service';
 import { AssetInfoRepo } from '@/features/assets/asset-info.repo';
 import { AssetRepo } from '@/features/assets/asset.repo';
 import { AssetsController } from '@/features/assets/assets.controller';
@@ -51,6 +52,7 @@ import { WorkgroupsController } from '@/features/workgroups/workgroups.controlle
     provideElasticsearch,
     AssetEditRepo,
     AssetInfoRepo,
+    AssetEditService,
     AssetRepo,
     AssetSearchService,
     AssetSyncService,
